Move options constants to the central constants file
The archid options values were being defined where they were used. This is not a good practice as it would mean that if the row number in the options file changed then there all files would need to be searched to find where it was defined.

The af_constants_h file has been included in af_option_h to ensure that the constant values are always available to any script calling the options wrapper script, which should be all scripts.

There is no need for any recompilation as the constant values have not changed, only moved location.

see #22