Merge 'Attack resolution' tweak from Dain's Fixes
Updates various components that are all linked to gether at the source code layer. The changes to the combat_h file were includes in a previous mod, but are enabled by this mod. Botrh the *_singletarget and the *_aoe_instant scripts were recompiled to pick up the changes in the included file, with no change the scripts themselves.

closes #27