Merge 'Blizzard disabling effects' tweak from Dain's Fixes
Existing behaviour only checks once each to slip and freeze - when entering the blizzard, and on the third (globally, not per unit) heartbeat respectively. Blizzard now has a 12.5% chance to do each of those actions on every heartbeat (75% chance of no effect). The freeze may also now allow enemies to be shattered.

closes #38