Merge 'Chain Lightning' fix from Dain's Fixes
Now benefits from spellpower, using the same scaling as other spells (i.e. 100 spellpower doubles damage). It also no longer bounces to the propagating target, can also continue to propagate if the bolt kills the target, and centres its AoE on the correct target when selecting enemies to bounce to.

closes #59