Add missing function definitions to toolkit
The toolkit gets the list of know functions from a file named script.ldf, which is a text file that is bundled inside an erf. Each patch contains an updated script.ldf file with the toolkit using the script.erf from the highest numbered erf that contains one.

The latest patch to contain a script.lst is patch3, however the lsf file it contains was very poorly generated and contains a huge number of typos and incorrect function comments that would show up in the toolkit help file. It also was missing multiple function definitions meaning some scripts would no longer compile.

The patch3 lsf contains no additional functions so it is safe to use the patch2 script.lsf file in the Documents\BioWare\Dragon Age\packages\core\patch folder to give the toolkit knowledge of all the functions and the correct help documentation.

This commit contains the file from patch2

closes #65