Combine existing area prcscr scripts
There were two scripts that are both run opn all areas and there is an expectation that there will be more.
It is inefficient to have to run multiple scripts running the same loops performing actions so these have been combined into a single script in this commit so the loops are only run once.

closes #70