Fix event manager script naming for order
The numbering of event types in event manager was misinterpreted whern it came to the event names.

The base event manager code from the original mod had 1 as the post handler and 2 as the pre handler. The handler contained within the Dain's Fixes mod, which is what is used here reversed this. This is more intuitive, but it meant that the file names used were confusing.

The correct code ran for each event as all the eventmanager values came form events in Dain's Fixes so they were consistent, the names of the scripts called however did not match.

The script names have been updated along with the corresponding rows in the 2da files, along with any coments indicating processing order

closes #80