Merge 'Spider Queen' tweak from Dain's Fixes
Fixes some cases where the Corrupted Spider Queen won't trigger web traps or summon reinforcements, and increases the damage of her Poison Spit talent by 50%. A 2x multiplier was originally implemented in game code but is commented out with the note "removed, too powerful". The DOT stacking fix applies a severe nerf to this ability however, so this tweak is to restore the vanilla difficulty of the fight.

closes #131