Fix issue with lucky effect
The code from Dain's mod that handles the Lucky item property has the potential to permanently reduce creature stats if the mod is enabled after an item with the effect is equipped by a creature.
This commit resolves this storing whether followers have received the effect in a module variable. A bit is set for each follower if the effect has been applied to them. The effect will only be applied if the follower does not currently have the effect applied, and vice-versa.
This is not tracked for non-followers, but it is considered very unlikely that the mod will be installed between the time an NPC has the effect applied via an item and the item being removed given that items are rarely removed from NPCs.

closes #140