Add 'Immersive DLC Integration' mod
Instead of dumping all DLC/promo items in your inventory or Bodahn's shop, distributes them throughout the game in lore appropriate spots, adding 57 'new' items to the looting experience. Includes a rebalance of the Feastday Gifts.
Includes the batch file in the sources/tools folder that patches the items for iinclusion in Awakening

closes #147