Add 'Awakening Silverite Mines Bugfix' mod
This mod fixes the infamous Silverite Mines equipment glitch that causes you to permanently lose some/all of the equipment on your player character. With this fix installed the scene involving the Architect & Utha will display correctly (not against a black backdrop) and your equipment will be correctly removed at the appropriate time.

The dialog file in the mod is superseded by the one in [Awakening Fixes and Tweaks](https://www.nexusmods.com/dragonage/mods/5944) which incorporates it plau adds additional fixes

closes #172