Add 'Sulcher's Pass consistency' mod
Adjusts Sulcher's Pass quest to start after you visit camp, gives you a new letter from Felix and fixes the error in Shale's codex.

Merged the components to start the mod from the players camp and the updated Shals codex entry. The component to add the letter to Bodahn in the camp has not been merged.

closes #180