Merge 'City Elf won't strip for King Cailan' mod
If Soris's wedding clothes were in the City Elf's possession when they left the alienage, they should remain in the City Elf's possession.
That way you won't accidentally meet the king naked if you put on Soris's clothes before you left.

There was no source code with the mod, but the change is trivial so updated the source code so other changes to the same file may be merged.

closes #183