Add 'Surana Dialogue Fixes' mod
In the mage origin, if you ask First Enchanter Irving to sign the form for the rod of fire, he’ll warn the soon-to-be Warden about the wards on the phylactery chamber door, and he’ll suggest a way around them. That’s supposed to allow the Warden to recognize the Tevinter artifact in the basement, but unfortunately, Irving’s dialogue is bugged. This is my attempt to fix it.

Mother Hannah Dialogue Fix: When you meet Mother Hannah in Redcliffe, she greets all elven Wardens in the same way, regardless of their class. I added a dialogue option that allows Surana to tell her, “I am also a mage.” (This unlocks a part of the conversation that was previously only available to human mages.)

closes #184