Add 'Ghost Elves' mod
There are three ghosts encountered in the Brecilian Ruins during Nature of the Beast- all of them speak Elvish, but the two who get actual conversations with the Warden (a mother and child) both use human character models. This could be intentional on Bioware's part, but I thought it was a little strange considering that they spoke Elvish and the toolset describes the child as elven. Accordingly, this mod makes the ghost boy and his mother elves while otherwise keeping their appearance the same as their vanilla counterparts.

closes #189