Add the 'TSM Werewolf Fix' mod
Cassian, the leader of the werewolf contingent who appear at Redcliffe Castle if the Warden sides with the werewolves during Nature of the Beast, does not have an eye mesh assigned to his model; a second werewolf in the company also suffers from this omission.

The commit adds the missing mesh assignment to their character files.

closes #194