Add 'More Detailed Tooltips' mod
It includes tooltips for DLC including The Stone Prisoner, Warden's Keep, Return to Ostagar, Dragon Age: Awakening, Golems of Amgarrak and Witch Hunt, item properties, item sets, and a large number of additions and corrections to Bibdy's work (including all of his To-Do section). It is also based off the tlk table for patch 1.04, both to keep it up to date and so that I'm technically starting from a fresh source rather than copying existing work.

closes #204