Add 'Clear Dreams - no blur in the Fade' mod
Removes uncomfortable postprocessing effects in all Fade scenes: the radial blur over the whole screen, and the wavy distortion at the edges. There are 2 options with the mod and this selects to significantly reduce the distortion rather than completely remove it.

The distortion was erroneous uploaded with the previous commit, but has been moved to a better location with this commit.

closes #216