Fix broken placeables
Placeable objects were not being added to the area maps

This was due to a typo in the placeable_types 2da file where there was no mod extension added to the file, overwriting the base array instead of updating. This commit adds the _af extension to the 2da filename to resolve this.

closes #219