Add 'Desire Demon HD Retexture Catachrism' mod
This is an adult, yet lore friendly version of the Desire demon. I tried to stay true to concept however I replaced part of the chain that covered the nipples with a piercing concept that I thought was more in line with a demonic S&M theme.

I created this texture with a mature audience in mind thinking along the lines of The Witcher series.

Other details about the texture: heavily reworked the chains, skin, and fabrics for visual appeal and realism.

closes #249