Add 'Skeletons of the Grey Wardens in Return to Ostagar'
This mod adds Awakening creatures to the Return to Ostagar area
- Armoured Ogre
- Blighted Werewolf
- Skeletons

It also changes the Ostagar area:
- Add key to wizards chest
- Add Runic Forge

The base mod includes all the Awakening resources such as the creature ability animations and scripting events, etc. These should be available via the /packages/core_ep1 folder without adding them as loose files so they shoud not be included with this commit.

It needs to be tested to confirm that their exclusion works.

see #257