Add 'Dragon Age Redesigned'
This mod will redesign the WHOLE of the NPCs in Dragon Age to fix face proportions in an attempt to make each character look a little bit more realistic in Dragon Age. A lot of time and thought was put into each character design for this project.

This mod makes changes to the NPCs within Dragon Age Origins and its expansions and DLCs. The follower morphs are not used.

closes #273