Add 'Templar Variety - DAO'
DAO templars are all the same. Literally, most of the templars in DAO are clones of the same few characters (and despite the fact that female templar robes exist in the files and are perfectly usable, they're all men until Awakening). So, I've ported the medium templar armors from DAI, and distributed them to give the templars more variety.

closes #283