Fix DLC Integration mod merge
The merge of the DLC Integration mod was not working as the items are still being placed into the player inventory at the start of the game.
The root cause is that the key to the functionality is that the prcscr sripts that come as part of the DLC need to be overwritten rather than extended. In order to do this it is necessary to create multiple prcsrc 2da files with the same names.

The original mod created these 2da files and placed custom scripts inside, however the strategy of the archid mod is to have a single area script where possible so in order to achieve this dummy porcscr 2da files that have null entries should be created to override the dlc 2da and prevent the addition of the items into the players inventory, whist retaining the existing shared area scripts.

closes #297