Add 'Goblet De-Gobbler - a Joining chalice fix and more'
The main function of this mod is to stop the chalice/goblet rotating weirdly when using some modded armour/clothing in the Joining cutscene in the base game. Vanilla and modded armour/clothing not affected by the issue will still behave correctly. It also fixes some other issues with the cutscene, and provides some options for replacement of Duncan's dagger.

closes #299