Add Gobblet De-Gobbler patch for Grey Wardens of Ferelden
The Goblet De-Gobbler mod has a patch for Grey Warden's of Ferelden to allow the GWoF to provide grey warden gear to the hero following the joining ceremony. There is no source for this patch so it cannot be merged. If there are subsequent changes to the plot script required a decision will need to be made whichj should eb kept, however given that tyhis is a plot script for the joining ceremony only that is unlikely.

closes #303