Add 'AlistairsRose'
This mod adds Alistair's Rose as an amulet, but it is not integrated with the game so a debug script is required to add it to the player inventory. The mod is being included as it forms part of a later quest mod.

The script to add the item manually should not be included

closes #313