Add 'Armor of the Sten' plus updates
The base Armor of the Sten mod adds a really cool armor set for Sten to the game. However, because the item is a piece of clothing, it does not offer any armor bonus or fatigue penalties. This mod author did add an armor bonus, however without the fatigue penalty, it causes the armor to be overpowered.

I have added a fatigue penalty to the armor, and a set bonus with Heavy Chainmail boot and gloves. The armor has been moved to the mercenary outside of Lothering, who previously carried the Heavy Chainmail armor (meant for Sten).

Allows Epilogue!Sten to wear the snazzy Armor of the Sten, rather than some random massive armour I never put on him.

closes #321