Add 'TSM The Unobtainables'
Most of the items added to Old Tegrin are gifts that can be found in the toolset, but not in the game.

- Armband (gen_im_gift_armband.uti) / According to the toolset, this item was added as loot to a corpse in the Brecilian Forest, but that corpse was not placed in the game.
- Black Gemmed Ring (gen_im_gift_ring4.uti) / Unobtainable by normal means, per the DA Wiki.
- Diamond Earrings (gen_im_gift_earring.uti) / Unobtainable by normal means, per the DA Wiki.
- Heavy Gold Ring (gen_im_gift_ring3.uti) / Unobtainable by normal means, per the DA Wiki.
- Botanist's Map of Thedas (gen_im_gift_map4.uti) / This gift is in a chest in Redcliffe Castle during the climax, but in an inaccessible room.
- Gateway Amulet (gen_im_acc_amu_amu8.uti) / Not a gift! According to the Wiki, this amulet is carried by an NPC in Lothering, and was probably intended to be available via stealing. The NPC in question cannot be selected, however, and therefore cannot be stolen from.

Both of the items added to Haven would have been available from a second shopkeeper, if the player killed the first merchant but later sided with the cult. That shop was apparently not implemented in the game, though, and thus the items were only available via the console. I've added one of each to the existing shopkeeper's inventory. Neither is lootable, and both are pricey.

- Greater Elixir of Grounding Recipe (gen_im_cft_hrb_406.uti)
- Greater Tome of the Mortal Vessel (gen_im_qck_book_attribute2.uti)

According to the DA Wiki, two of Shale's gifts are not available in the PC version of the game, and can only be obtained via the developer console. I added both to Garin's inventory.

- Remarkable Garnet (gem_im_gift_gar.uti)
- Remarkable Diamond (gen_im_gift_dia.uti)

- North Ward (gen_im_acc_amu_men20.uti) / Not a gift, and not available in-game without the developer console, per the DA Wiki. Two different darkspawn carry this amulet during the climax, but is not lootable from either. I set it to be dropped as loot by one of them (the Hurlock General in the Alienage).

closes #327