Fix areas not having their prcscr scripts being run
The change to use a bit flag to ensure that prcscr scripts for an area list were only run one introduced a bug where there are multiple areas per area list. In this scenario the flag would be set onm the first area entered and scripts that run against other areas would never be fired.

This commit has implemented a bit flag for each individual area and also updated the header comment of each script to clearly indicate the list of areas each script covers

closes #330