Add 'Skip the Fade'
When you defeat Duncan and use the pedestal, you will be teleported to each of your companion's nightmares. If all companions are rescued, just click on the pedestal and you will be teleported to the Sloth Demon. From there on all will proceed as usual: kill the Sloth Demon, Neil appears and you are teleported back to the Circle Tower.

You get all the attribute points, codex entries and XP you normally would get, when you complete the Fade.

closes #356