Add 'DahliaLynn's Sleep Until Dawn' patch
Click on your camp tent to go to sleep alone, with dog, or snuggle with a love interest. Wake up to a beautiful camp morning.

The core mod must be installed using the dazip as it contains a module script with no source so cannot be merged. There is a patch, however that is installed in the override folder, which is included in this commit.

closes #359