Add 'The Rescue at Ishal'
This mod extends the ending of the battle of Ostagar (mod contains spoilers). It adds two new cutscenes and two new pre-filmed movies, showing details of the devastating aftermath of the battle, what happened to Alistair and the Warden after the tower ambush. Lore-friendly. For all races, genders, and origins-NOT recommended for first-time players.

The 4k bik video files are not included due to github file size restrictions.

closes #361