Add 'Spiders and Knives' quest
The mod takes place in and around Lothering, so the initial quests need to be completed before Lothering is destroyed by the darkspawn. The mod is suitable for any class - the main reward item at the end of the main path is tailored to the main player character.

As far as I can tell, the difficulty level is broadly similar to other Lothering quests, whilst the rewards are a little higher - although hopefully not to an unbalancing extent. The main quest path has around 40 minutes of gameplay; there is an alternative, shorter, path of around 15-20 minutes which also provides a nice item as a reward, but less gold.

To begin the quest, look at the Chanter's Board in Lothering. You might also stumble upon the starting point of the quest whilst exploring around Lothering.

The initial quest phases end once Knife Edge is secure (if you take the main path). If you then return after the fall of Lothering there is more to do.

The core files have been placed into the override folder and shared gda files have been merged into the existing 2da files.

This issue is not being closed with this commit as there is no character far enough into the game to perform testing

see #362