Add 'Extra Dog Slot'
This mod allows the dog to tag along with the PC for most of the game as a
fifth companion residing in the top-most available summon slot. (summon slots
are mini-portraits that reside to the right of each NPC portrait.)

When presented with a party selection screen, if you do not choose the dog as
one of your 4 primary companions, he will automatically appear next to the PC
as a fifth companion after the selection screen goes away.

When present as a 5th party member, you can tell the dog to leave or call him
back by using a new Dog Whistle item which will appear in your inventory after
you aquire the dog as a companion.

While the mini-slot is normally used for summoned creatures, this mod manages
the slot so that the dog is a full blown party member and not a simple summon.
All dialog options, banter and special events and abilities involving the dog
will still work.

see #364