Add 'Inquisition Fade'
The rocks are 2k textures now, and I used a preexisting texture, made it green and overlaid it over all of them to give off the same vibe as the ones in DAI. Main reason I made the rocks 2k is because for some models they're kinda stretched (which you can see in one of the screenshots still) and I tried to make it look less bad. Also the rocks are shiny now.
Changed the terrain a bit and made it darker to match the rocks.

coses #372