Fix area scripts run for every area in area list on first entry
Some PRCSCR spripts run against an area list rather than individual areas and for these if there is code that should only be run once the script uses a bitflag to wrap the code. This does not work as expected for two reasons:

- The flag conditions are not further wrapped in a condition that checks the actual area. This means that the code for every area will be run and teh flags set to completed on the first area being visited
- The use of the flags is not a good mechanism in some cases as these mean that the script cannot be updated to make additional changes once an area has been visited. For some use cases the single use is fine, but for others such as adding immovable objects it is better to check for the existence of the item. Items that may be sold should still use the bit flag to prevent a player collecting the item, selling it and then going back to the area for another.

In all cases a wrapper condition that checks the current area tag has been used, even for single area scripts, for consistency.

closes #377