Give Arcane Warrior armour set on releasing spirit in brecilian ruins
In addition to unlocking the Arcane Warrior class, releasing the anciemt arcane warrior spirit now also provides the player with a set of armour and weapons. The rewards are based on The Ivory Tower and Nightfall Bloom mod with updated stats. The original mod had different armour for male and female but this has been fixed so a single set may be worn by either sex.

see #381