Add 'Legacy of Sundown'
Full set of a light armor of elven craft for rogues made in a gold-orange-olive coloring with a marching set of two daggers two swords and a bow.
I have modified the stats to fit better into the power curve of my game.

These have been placed into the chest in the West Brecilian forest that appears if you defeat the Greater Shade at the ensorcelled camp.

see #384