Add 'The Phoenix Armory for Females'
A Full Set of Weapons and Armors for Female Rogues. This Project was inspired by the Fiery Phoenix of the Far East, the Matriarch of birds and couterpart to the Dragon.

The models are used unchanged but the item stats have been reworked to suit the power curve of this mod.
The armour and weapons are integrated into the game in various locations, either on enemies, in caches or in stores. The arrows may be bought from various stores, but are limited in number. The limit is based on the number of arrows that the party have in stock and merchants will replenish their supplies if the party use their stock.

closes #386 #387 #388 #389
see #385