Add 'Armor of Garahel'
This mod contains the armor and weapons of the legendary Grey Warden Garahel who ended the Fourth Blight.

A massive armour and sword set suitable for tanks if an appropriate shiled is found. The mod contains twin longswords whose stats have been modified to have one suitable for general enemies and one that is geared towards fighting archdemons.

see #390