Increase number of tactics slots
With the Even More Advanced Tactics mod the base numbver of tactics slots is not sufficent as the mod uses multiple slots to determine an enemy to select for skills and it is usual to have several multi-slot conditions so the normal progression is not really appropriate. The leveling table has been updated to add more slots each level.

closes #398