Make Legacy of Sundown set more generic
The Legacy of Sundown set was too targetted at a specific cunning based rogue build. It should be more generic allowing it to be used across multiple builds instead. This commit acheives this by updating a few of the items to increase dexterity rather than cunning in the item properties as this attribute is useful to more builds.

see #384